+++ client/src/Components/_example_LargeMultiComponent/index.js
@@ -1,10 +1,12 @@
 import React from 'react';
-import TestComponent  from './testComponent.js';
+import TestComponent  from '../_example_testComponent.js';
+import './style.css'
 
+/* This component has unique style */
 export default class LMC extends React.Component {
   render() {
     return (
-      <div>
+      <div id="l-m-c">
           <strong>Using the Test Component: </strong><TestComponent />
         <div>
           <strong>Using the Test Component again: </strong><TestComponent name="nemo"/>
