+++ client/src/Components/_example_LargeMultiComponent/style.css
@@ -1,0 +1,3 @@
+#l-m-c {
+  background-color: grey;
+}
