+++ client/src/Components/index.js
@@ -1,8 +1,8 @@
 /* This file will be used to combine all components into a single library that
  * we will that reference in App.js*/
 
-import TestComponent  from './testComponent.js';
-import LMC  from './largeMultiComponent.js';
+import TestComponent  from './_example_testComponent.js';
+import LMC  from './_example_LargeMultiComponent/index.js';
 
 
 export {TestComponent, LMC}
